refactor: Extract variant analysis tools into separate module
- Create src/tools/variant_tools.py with 13 variant functions
- Move variant analysis logic from server.py:
  - get_variant_dbnsfp()
  - get_clinvar_by_variant()
  - get_clinvar_by_gene_symbol()
  - get_clinvar_by_entrez_id()
  - get_gnomad_variant()
  - get_gnomad_by_gene_symbol()
  - get_gnomad_by_entrez_id()
  - get_dgv_variant()
  - get_dgv_by_entrez_id()
  - get_decipher_variant()
  - get_decipher_by_location()
  - get_geno2mp_variant()
  - get_geno2mp_by_entrez_id()
- Add register_tools() function for clean registration
- Update server.py to import and register variant tools

Benefits:
- Isolates complex variant analysis logic (largest module)
- Reduced server.py by ~280 lines
- Easier to maintain database-specific integrations
- Better code organization

All existing functionality preserved - backward compatible.
